dynamic channel model cleanup

+++ gr-channels/lib/cfo_model_impl.cc
@@ -68,7 +68,7 @@ namespace gr {
 
         const gr_complex* in = (const gr_complex*) input_items[0];
         gr_complex* out = (gr_complex*) output_items[0];
-        for(size_t i=0; i<noutput_items; i++){
+        for(int i=0; i<noutput_items; i++){
             // update and bound cfo
             // we multiply by a random {1,-1} to remove any sign
             // bias that may exist in our random sample pool
